Moved all Deprecated tests to own file
